fix: comprehensive 64-bit architecture audit and modernization
- Upgrade marshall_poiner in type.h to use uintptr_t instead of uint32_t
- Support 64-bit size_t allocation in CreateFileMapping (sharemm.h)
- Fix socket non-blocking ioctlsocket parameter type in tcpsocket.h
- Fix 64-bit registers pass-by-reference in fptrace.h
- Update SAFE_TEST_INDEX in macro.h with 64-bit integer casts
- Fix string buffer underflow and bounds checks in fstring.h
- Update fallocapp test loops to use size_t, fixing 64-bit sign-extension warnings
- Modernize tools/genstat with Go module, rebuild 64-bit ELF and PE32+ binaries
- Remove legacy 32-bit MFC watcher application and binaries

+++ fakengine/type.h
@@ -40,7 +40,7 @@ union marshall_float
 
 union marshall_poiner
 {
-	uint32_t i;
+	uintptr_t i;
 	void * p;
 };
